move Updater and friends to ext submodule
